requirements.txt
Flask==3.0.3
google.generativeai==0.8.3
werkzeug==3.1.3
PyPDF2==3.0.1
pdfplumber==0.11.4
flask_cors==5.0.0
python-dotenv==1.0.1